Update Lab9_Act3.py
I have completed part A & part B. Both parts output the correct data, but the formatting on the only first row is misaligned. Josh, this is the only thing that needs fixing before completing part C. All of the data categories (temperature, precipitation, etc.) have been stored in their own lists so this should make using and formatting the data a bit easier.